Move parse_res_to_dict to requests.py
